Add NCP car parks for UK
NCP (National Car Parks) are probably the major operator of commercial car parks across the UK. Their yellow logo is very recognisable, and around 170 are already tagged with operator=NCP on OSM, with a smaller number tagged with name=NCP. Most car parks will have a distinct name because they will operate several car parks in most city centres, and therefore pre-populating the name tag is likely to be wrong, but operator is correct. 

I have added a matchNames item, but am not sure if this will work with operator, and I'm afraid I have to leave this to the maintainers' judgement.